feat: :zap: F64 conversion

+++ assembly/UniswapV3LiquidityStrategy.ts
@@ -1,11 +1,11 @@
 import { BaseStrategy } from '@steerprotocol/base-strategy/assembly';
 
 export class UniswapV3LiquidityStrategy extends BaseStrategy {
-  private LIQUIDITY_PRECISION: f32 = 2 ** 16 - 2; 
-  private startTick: f32;
-  public binWidth: f32 = 600;
+  private LIQUIDITY_PRECISION: f64 = 2 ** 16 - 2; 
+  private startTick: f64;
+  public binWidth: f64 = 600;
 
-  constructor(binWidth: f32) {
+  constructor(binWidth: f64) {
     super();
     this.binWidth = binWidth;
   }
